refactor(web): 단일 도메인 전용 api를 domain/{X}/api로 이동
vertical(DDD) 구조에 맞춰 단일 도메인에서만 쓰는 api 클라이언트를 해당 도메인으로 이동:
- domain/challenge/api: attendance, attendanceSchema, challengeQuestion, presignedUrl
- domain/faq/api: faq
- domain/mentor/api: mentor, challengeMentorGuide(+Schema, +test)
importer 경로 일괄 갱신. 공유 의존(challengeSchema, userExperienceSchema, mentorSchema)은 api/ 유지.
typecheck/lint/prettier 통과, 동작 변경 없음.
