+++ apps/web/src/domain/challenge/api/attendanceSchema.ts
