+++ apps/web/src/domain/challenge/my-challenge/section/DailyMissionSubmitSection.tsx
@@ -1,4 +1,4 @@
-import { usePatchAttendance } from '@/api/attendance/attendance';
+import { usePatchAttendance } from '@/domain/challenge/api/attendance';
 import {
   useGetChallengeReviewStatus,
   usePostChallengeAttendance,
