+++ apps/web/src/domain/challenge/my-challenge/section/mission-submit-bonus/useMissionSubmitBonus.ts
@@ -1,4 +1,4 @@
-import { usePatchAttendance } from '@/api/attendance/attendance';
+import { usePatchAttendance } from '@/domain/challenge/api/attendance';
 import { useChallengeMissionAttendanceInfoQuery } from '@/api/challenge/challenge';
 import { useSubmitMissionBlogBonus } from '@/api/mission/mission';
 import { useCurrentChallenge } from '@/context/CurrentChallengeProvider';
