+++ apps/web/src/domain/challenge/ot-mission/OtMissionSubmitMenu.tsx
@@ -1,4 +1,4 @@
-import { usePatchAttendance } from '@/api/attendance/attendance';
+import { usePatchAttendance } from '@/domain/challenge/api/attendance';
 import {
   useGetChallengeGoal,
   usePatchChallengeGoal,
