+++ apps/web/src/domain/mentor/api/challengeMentorGuideSchema.ts
